pyproject.toml
[tool.poetry]
name = "cassette"
version = "0.1.1"
description = ""
authors = ["Ole-Batting <batting.ole@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.8.1"
Pygments = "^2.17.2"
pyyaml = "^6.0.1"
numpy = "^1.26.2"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"